% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_download_data.R
\name{get_download_data}
\alias{get_download_data}
\title{get_download_data}
\usage{
get_download_data(current_tab, values, results_tab)
}
\arguments{
\item{x}{score}

\item{perc}{percent}
}
\value{

}
\description{
get_download_data
}
