% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA_get_norms.R
\name{get_norms}
\alias{get_norms}
\title{Get Norms}
\usage{
get_norms(stimulus, google_sheets = F)
}
\arguments{
\item{stimulus}{which stimulus was chosen}
}
\value{

}
\description{
Note that if googlesheets fails, static norms will be used.
}
