% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA_get_plot.R
\name{get_plot}
\alias{get_plot}
\title{get plot}
\usage{
get_plot(norms, current_score, stim)
}
\arguments{
\item{norms}{norms pulled from googlesheets or static if not accessed}

\item{current_score}{the total score}

\item{stim}{which stimulus was scored}
}
\value{
a density plot
}
\description{
get plot
}
