% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_report.R
\name{get_plot_report}
\alias{get_plot_report}
\title{get plot}
\usage{
get_plot_report(
  norms,
  current_score,
  stim,
  scoring = "dalton_richardson",
  norm_var,
  basesize
)
}
\arguments{
\item{norms}{norms pulled from googlesheets or static if not accessed}

\item{current_score}{the total score}

\item{stim}{which stimulus was scored}

\item{scoring}{scoring}

\item{norm_var}{norm_var}

\item{basesize}{base size of plot}
}
\value{
a density plot
}
\description{
function to build the plot for the results page
}
