% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_text.R
\name{get_results_text}
\alias{get_results_text}
\title{Generates results text for top of results page}
\usage{
get_results_text(results, num_concepts, time)
}
\arguments{
\item{results}{results_mca_tab()}

\item{values}{values reactive}
}
\value{
some text to be put in a p() tag.
}
\description{
Generates results text for top of results page
}
