% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_scoring.R
\name{get_scoring_div}
\alias{get_scoring_div}
\title{Get scoring Div}
\usage{
get_scoring_div(num)
}
\arguments{
\item{num}{current item to be shown}
}
\value{

}
\description{
This function renders the UI for input$mainpage == "intro" on the navbar.
}
