% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA_get_sentence_data.R
\name{get_sentence_data}
\alias{get_sentence_data}
\title{Get sentence data}
\usage{
get_sentence_data(rds_list)
}
\arguments{
\item{rds_list}{a list sentences}
}
\value{

}
\description{
Get sentence data
}
