% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA_get_summary_table.R
\name{get_summary_table}
\alias{get_summary_table}
\title{Get summary table}
\usage{
get_summary_table(results, norms, scoring = "dalton_richardson", min = 0)
}
\arguments{
\item{results}{the data frame of results}

\item{norms}{the norms for the scored stimulus}
}
\value{

}
\description{
Get summary table
}
