% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_table.R
\name{get_training_table}
\alias{get_training_table}
\title{get training table function}
\usage{
get_training_table(dat_in)
}
\arguments{
\item{values}{}
}
\value{

}
\description{
get training table function
}
