% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_trascript_div.R
\name{get_training_trascript_div}
\alias{get_training_trascript_div}
\title{get training transcript}
\usage{
get_training_trascript_div(module_num)
}
\arguments{
\item{module_num}{}
}
\description{
get training transcript
}
