% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{main_concepts}
\alias{main_concepts}
\title{Main Concept Data}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 72 rows and 6 columns.
}
\usage{
data(main_concepts)
}
\description{
Main Concept Data
}
\keyword{datasets}
