% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mainConcept.R
\name{run_mainConcept}
\alias{run_mainConcept}
\title{Run Main Concept Function}
\usage{
run_mainConcept()
}
\value{
a shiny app!
}
\description{
The main function of the package to open the app
}
\examples{
\dontrun{
run_mainConcept()
}
}
