% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{static_norms}
\alias{static_norms}
\title{Normative Data - static}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2045 rows and 10 columns.
}
\usage{
data(static_norms)
}
\description{
Normative Data - static
}
\keyword{datasets}
