% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_download_data.R
\name{get_download_data}
\alias{get_download_data}
\title{get_download_data}
\usage{
get_download_data(current_tab, values, results_tab)
}
\arguments{
\item{current_tab}{current navbar tab selected}

\item{values}{all the values and stuff}

\item{results_tab}{the table from the results page}
}
\value{
a list of data that goes into writeexcel (list elements to tabs)
}
\description{
organizes data for downloading
}
