% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long_results_df.R
\name{get_long_results_df}
\alias{get_long_results_df}
\title{Get long results data}
\usage{
get_long_results_df(concept_accuracy, filtered_concepts)
}
\arguments{
\item{concept_accuracy}{current concept scores}

\item{filtered_concepts}{data of concepts relevant to the stimulus}
}
\value{
a long dataframe of results
}
\description{
gets long results for calculating stuff and for downloading data.
THIS IS A HUGELY IMPORTANT FUNCTION FOR CALCULATING THE RESULTS!!!
}
