% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_norms.R
\name{get_norms}
\alias{get_norms}
\title{Get Norms}
\usage{
get_norms(stimulus, google_sheets = F)
}
\arguments{
\item{stimulus}{which stimulus was chosen}

\item{google_sheets}{should the functionget norms from google sheets?}
}
\value{

}
\description{
Note that if googlesheets fails, static norms will be used.
}
\details{
gets normative data from google sheets, or just uses a locally stored
version that has to be updated manually
}
