% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentence_data.R
\name{get_sentence_data}
\alias{get_sentence_data}
\title{Get sentence data}
\usage{
get_sentence_data(rds_list)
}
\arguments{
\item{rds_list}{a list sentences}
}
\description{
Not currently used as far as I can tell
But not deleting it until I confirm that for sure
I suspect this was made early on when I was trying to figure out
The best way to deal with the incoming text data...
}
